refact runner model v1

refact model runner v1
